+++ server.crt
@@ -1,0 +1,20 @@
+-----BEGIN CERTIFICATE-----
+MIIDOjCCAiICCQCqBXpsQm3usjANBgkqhkiG9w0BAQsFADBfMQswCQYDVQQGEwJE
+RTEPMA0GA1UECAwGQmF5ZXJuMQ8wDQYDVQQHDAZNdW5pY2gxCzAJBgNVBAoMAlZP
+MSEwHwYJKoZIhvcNAQkBFhJwb3N0QGp1cmVrYmFydGguZGUwHhcNMTgwNTA0MTIy
+NjUwWhcNMjgwNTAxMTIyNjUwWjBfMQswCQYDVQQGEwJERTEPMA0GA1UECAwGQmF5
+ZXJuMQ8wDQYDVQQHDAZNdW5pY2gxCzAJBgNVBAoMAlZPMSEwHwYJKoZIhvcNAQkB
+FhJwb3N0QGp1cmVrYmFydGguZGUwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEK
+AoIBAQDZQcNQB8p5hEkbmwZaK2OcrsOEXLMYe0H/TXz78zkeEzadVPIax1sLCs2D
+vT1SVIAvaZqzYkBpkRlg8guCdQDIEkXE2+KRmVPtKyOVUinSNNBXmzVjER3FUoG7
+soogW3WsQfKTAGIiYzBpTe5c2LwPlLbcGYOG87cxuHhRJdaWbZbwf84vQNRFY52I
+xLhGQ2jxMrNrXAqdaWNtkNqFFEMBV9RVA5vtUBHV/o4XMnVyRapC0DOkAahrYxAi
+CtzhuB/cM8cuQ/lUI/smUG9ytkzMm+SxGdhVeUtkNOhvu8PL8gwak/po0BPdQcVa
+PBpcHy/Dn2pysKaN7elYFeIZ2TJbAgMBAAEwDQYJKoZIhvcNAQELBQADggEBACvF
+oV4hV/uk87rmIcINQIkF6KxsWSYmprXltYFAWX4uiqLWIQIvnMvIJIDdCSduQuq+
+rz3PkEX/SO/4sYGKa9A4gQLgU3EK0XUqd98gahINJZhRc6k7CdhpIytzYlRcSogh
+5GPg7f+esZiO1qEM4X89OqBIQbMUftwpMwMlz1Wpf6YpBSAkj0s5FrMipUmRlN5I
+uUqr8gkXWowdt4xUWE8eA1PhzYU48CONki3/bWw9NZmlURYcRIO3ArpDqU7KJNBz
+0LcIECCJAxJ+//K9IPyeXURcpW72TC1B83bGk8Lj6/AxFpX6IwiKBqwbrKXwZK9W
+2YLdxTgpMFBjHSKIIuI=
+-----END CERTIFICATE-----
