+++ src/getResourcesFromDom.js
@@ -1,7 +1,80 @@
-import deepEqual from './deepEqual';
 import produce from 'immer';
 import random from './randomId';
-import getPath from './getPath';
+
+function getOrigin() {
+  if (!window.location.origin) {
+    return (window.location.origin =
+			window.location.protocol +
+			'//' +
+			window.location.hostname +
+			(window.location.port ? ':' + window.location.port : ''));
+  }
+  return window.location.origin;
+}
+
+function absolutePath(urlString) {
+  let normalizedUrl;
+  const absoluteReg = new RegExp('^(?:[a-z]+:)?//', 'i');
+
+  if (absoluteReg.test(urlString)) {
+    // is absolute
+    normalizedUrl = urlString;
+  } else {
+    if (urlString.indexOf('/') === 0 && urlString.indexOf('//') !== 0) {
+      normalizedUrl = getOrigin() + urlString;
+    } else {
+      const paths = window.location.pathname.split('/');
+      paths.pop();
+      normalizedUrl = getOrigin() + paths.join('/') + '/' + urlString;
+    }
+  }
+  return normalizedUrl;
+}
+
+function getPath(path, resource, options) {
+  const pathIsRelative =
+		path.indexOf('http://') === -1 &&
+		path.indexOf('https://') === -1 &&
+		path.charAt(0) !== '/';
+
+  // resolve relative paths
+  if (pathIsRelative) {
+    if (resource.base) {
+      if (resource.base.substring(0, 2) === '##')
+        resource.base = options.baseMap[resource.base];
+
+      path = resource.base + path;
+    } else if (options && options.base) {
+      path = options.base + path;
+    }
+  }
+  // create unique absolute path
+  const normalizedPath = absolutePath(path);
+
+  return normalizedPath;
+}
+
+// not really deep equal, more tailored to the needs of resourceLoader
+const deepEqual = (a, b) => {
+  if (a.length !== b.length) return false;
+  for (const key in a) {
+    const aa = a[key];
+    const bb = b[key];
+
+    if (
+      aa.paths.length !== bb.paths.length &&
+			aa.dependsOn.length !== bb.dependsOn.length
+    )
+      return false;
+    for (const dependsKey in aa.dependsOn) {
+      if (aa.dependsOn[dependsKey] !== bb.dependsOn[dependsKey]) return false;
+    }
+    for (const pathKey in aa.paths) {
+      if (aa.paths[pathKey] !== bb.paths[pathKey]) return false;
+    }
+  }
+  return true;
+};
 
 const getResourcesFromDOM = (selector, options, resources = []) => {
   // Array of HTML elements with a dataset "resources"
